requirements.txt
customtkinter
pygame
pystray
Pillow
requests
beautifulsoup4
urllib3